Move shape components to physics.components.shapes
