+++ engine/src/main/java/org/terasology/physics/components/shapes/HullShapeComponent.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package org.terasology.physics.shapes;
+package org.terasology.physics.components.shapes;
 
 import org.terasology.entitySystem.Component;
+import org.terasology.rendering.assets.mesh.Mesh;
 
 /**
  */
-public class SphereShapeComponent implements Component {
-    public float radius = 0.5f;
+public class HullShapeComponent implements Component {
+    public Mesh sourceMesh;
 }
